refactor: extract shared command response formatting
